requirements.txt
flask==2.2.5
redis==4.5.5
pyjwt==2.6.0
